txt2graph/requirements.txt
# Zep Cloud SDK
zep-cloud>=2.0.0

# PDF处理
PyMuPDF>=1.24.0

# Markdown处理
markdown>=3.5.0

# Web可视化界面
streamlit>=1.38.0

# 图可视化
pyvis>=0.3.2

# 环境变量
python-dotenv>=1.0.0

